Completed the algorithm. I think it works ok now !!
The only issue that I really can't figure out how to deal with is when the maximum number of team members is greater than or equal to the counts of people who chose any option. This shouldn't be a problem with many users and/or when there is a few options to choose, but could be an issue when theres a small amount of users and/or alot of options to choose from.